requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
GeoAlchemy2==0.14.1
psycopg2-binary==2.9.7
SQLAlchemy==2.0.20
pytest==7.4.2